requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
streamlit>=1.37.0
pydantic>=2.8.0